WICKET-4707 - IJavaScriptLibrarySettings to manage HeaderItems
Ultimately, IJavaScriptLibrarySettings is responsible for making sure
that the correct header contributions are made to allow JavaScript
components to work. This should be done with HeaderItems, which is the
correct abstractin of the problem. Simply wrapping ResourceReferences in
HeaderItems throughout the code makes for a lot of duplication.

This is part of my effort to remove ResourceReference.getDependencies()